Add files via upload
I have the code working to get the posts from firestore. It is formatted. I decided to get rid rid of the 4 column and just go with one to make things easier. I added a color behind the feed. But do not get alarmed, this color should not be definite. I just added to add whatever color just for now. 

The name in the post is just hard coded right now. We will have to add code to get the name and email from the database. 

Let me know if you have any questions.